chore: make the config simplier (#712)
