+++ zmon-controller-app/src/main/resources/config/application-github.yml
@@ -24,6 +24,6 @@ zmon:
           - firstTeam
           - secondTeam
           - thirdTeam
-    rest:
-      get-alert-results:
-        allowed-filters: [ application ]
+
+  alert-results:
+    allowed-filters: [ application ]
