+++ zmon-controller-app/src/main/resources/config/application-local-test.yml
@@ -41,6 +41,6 @@ zmon:
           - firstTeam
           - secondTeam
           - thirdTeam
-    rest:
-      get-alert-results:
-        allowed-filters: [ application ]
+
+  alert-results:
+    allowed-filters: [ application ]
